Move testing classes out of main file and refactor utils
